+++ packages/insomnia/src/ui/components/sidebar/sidebar-filter.tsx
@@ -1,19 +1,26 @@
 import React, { FC, useCallback, useRef } from 'react';
+import { useSelector } from 'react-redux';
 
 import { SortOrder } from '../../../common/constants';
+import { database as db } from '../../../common/database';
 import { hotKeyRefs } from '../../../common/hotkeys';
 import { executeHotKey } from '../../../common/hotkeys-listener';
+import { sortMethodMap } from '../../../common/sorting';
+import * as models from '../../../models';
+import { isRequestGroup } from '../../../models/request-group';
+import { selectActiveWorkspace } from '../../redux/selectors';
 import { KeydownBinder } from '../keydown-binder';
 import { SidebarCreateDropdown } from './sidebar-create-dropdown';
 import { SidebarSortDropdown } from './sidebar-sort-dropdown';
 
 interface Props {
   onChange: (value: string) => Promise<void>;
-  sidebarSort: (sortOrder: SortOrder) => void;
   filter: string;
 }
-export const SidebarFilter: FC<Props> = ({ filter, sidebarSort, onChange }) => {
+export const SidebarFilter: FC<Props> = ({ filter, onChange }) => {
   const inputRef = useRef<HTMLInputElement>(null);
+  const activeWorkspace = useSelector(selectActiveWorkspace);
+
   const handleClearFilter = useCallback(() => {
     onChange('');
     if (inputRef.current) {
@@ -29,6 +36,31 @@ export const SidebarFilter: FC<Props> = ({ filter, sidebarSort, onChange }) => {
       inputRef.current?.focus();
     });
   }, []);
+  const sortSidebar = async (order: SortOrder, parentId?: string) => {
+    let flushId: number | undefined;
+    if (!activeWorkspace) {
+      return;
+    }
+    if (!parentId) {
+      parentId = activeWorkspace._id;
+      flushId = await db.bufferChanges();
+    }
+    const docs = [
+      ...(await models.requestGroup.findByParentId(parentId)),
+      ...(await models.request.findByParentId(parentId)),
+      ...(await models.grpcRequest.findByParentId(parentId)),
+    ].sort(sortMethodMap[order]);
+    // @ts-expect-error -- TSCONVERSION the fetched model will only ever be a RequestGroup, Request, or GrpcRequest
+    // Which all have the .update method. How do we better filter types?
+    await Promise.all(docs.map((doc, i) => models.getModel(doc.type)?.update(doc, { metaSortKey: i * 100 })));
+    // sort RequestGroups recursively
+    await Promise.all(docs.filter(isRequestGroup).map(g => sortSidebar(order, g._id)));
+
+    if (flushId) {
+      await db.flushChanges(flushId);
+    }
+  };
+
   return (
     <KeydownBinder onKeydown={handleKeydown}>
       <div className="sidebar__filter">
@@ -46,7 +78,7 @@ export const SidebarFilter: FC<Props> = ({ filter, sidebarSort, onChange }) => {
             </button>
           )}
         </div>
-        <SidebarSortDropdown handleSort={sidebarSort} />
+        <SidebarSortDropdown handleSort={sortSidebar} />
         <SidebarCreateDropdown />
       </div>
     </KeydownBinder>
